+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog #
 
+## v0.2 (2013/05/12) ##
+- Changed the structure and use separated command files instead of adding everything in side the setting file
+- removed the camelCase command for now and will put it back once I find a better algrothem
+- Add "apply to" view panel after choosing the command
+- Add recent commands log
+
 ## v0.1 (2013/05/06) ##
 - Added basic structure of the plugin
 - Added set parser command